backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
cryptography>=41.0.0
openai>=1.0.0
requests>=2.31.0
qdrant-client>=1.7.0
PyPDF2>=3.0.0

